Added bootstrap layout
included are the bootstrap files, plus the updated webpages using
bootstrap. also some of the python files were updated for extra
pages/displaying more information from the database on the website. when
you fork it and upload it to your computer to play around with, make
sure to update the templates URL